Don't use native mac screenshot logic on Sequoia or later
CGDisplayCreateImageForRect is deprecated starting from macOS 15.0
onwards, so use wx's built-in color picker there for the time being.
It won't actually work, but it'll at least compile, so this is better
than nothing. In the future, a new implementation using ScreenCaptureKit
can be added. Or the color picker can be reworked entirely, seeing as
at this point it's broken on pretty much every modern platform that's not
Windows.

By setting MACOSX_DEPLOYMENT_TARGET to 14.0 or earlier, binaries using
CGDisplayCreateImageForRect can be compiled even on Sequoia (or later),
and the color picker will still work on those.